Delete install.sh
there are several problems with install.sh
apt stdout is directed to /dev/null but stderr remains. you should been using &>/dev/null instead, but that's a minor problem.
the real problem is i'm unsure that installing packages in this way is correct and the way it links the script from repo instead of copying it.
the script also has no `--help` option. 

So i'd say, creating a `configure.sh` instead of `install.sh` that also has mac address and ssh key management that a user always can use it to configure the script would be better and more useful. So for now, let's delete this